+++ plugins/time-saver/src/components/TimeSaverPageComponent/TimeSaverPageComponent.tsx
@@ -17,11 +17,9 @@ import React, { useState } from 'react';
 import { Typography, Grid, Tabs, Tab, Divider, Paper } from '@material-ui/core';
 import {
   InfoCard,
-  Header,
   Page,
   Content,
   ContentHeader,
-  HeaderLabel,
   SupportButton,
 } from '@backstage/core-components';
 import { LocalizationProvider } from '@mui/x-date-pickers';
@@ -46,8 +44,18 @@ import {
   DateFiltersComponent,
   IFilterDates,
 } from '../DateFiltersComponent/DateFiltersComponent';
+import CustomHeader, {
+  HeaderProps,
+} from '../TimeSaverHeader/TimeSaverHeaderComponent';
+import { configApiRef, useApi } from '@backstage/core-plugin-api';
 
-const GaugesContainer = ({ dates }: { dates: IFilterDates }) => (
+const GaugesContainer = ({
+  dates,
+  hoursPerDay,
+}: {
+  dates: IFilterDates;
+  hoursPerDay: number;
+}) => (
   <Grid
     container
     spacing={4}
@@ -71,7 +79,11 @@ const GaugesContainer = ({ dates }: { dates: IFilterDates }) => (
     </Grid>
     <Grid item xs={6} sm={6} md={2}>
       <Paper elevation={0}>
-        <TimeSavedGauge number={8} heading="Time Saved [days]" dates={dates} />
+        <TimeSavedGauge
+          number={hoursPerDay}
+          heading="Time Saved [days]"
+          dates={dates}
+        />
       </Paper>
     </Grid>
     <Grid item xs={6} sm={6} md={2}>
@@ -87,7 +99,11 @@ const GaugesContainer = ({ dates }: { dates: IFilterDates }) => (
   </Grid>
 );
 
-export const TimeSaverPageComponent = () => {
+export const TimeSaverPageComponent = (props: HeaderProps) => {
+  const configApi = useApi(configApiRef);
+  const hoursPerDay =
+    configApi.getOptionalNumber('ts.frontend.table.hoursPerDay') ?? 8;
+
   const [selectedTab, setSelectedTab] = useState(0);
   const [selectedTeam, setSelectedTeam] = useState('');
   const [template, setTemplate] = useState('');
@@ -115,13 +131,11 @@ export const TimeSaverPageComponent = () => {
   return (
     <LocalizationProvider dateAdapter={AdapterLuxon}>
       <Page themeId="tool">
-        <Header
-          title="Backstage TS plugin!"
-          subtitle="Check saved time with TS plugin!"
-        >
-          <HeaderLabel label="Owner" value="Rackspace" />
-          <HeaderLabel label="Lifecycle" value="experimental" />
-        </Header>
+        <CustomHeader
+          title={props.title}
+          subtitle={props.subtitle}
+          headerLabel={props.headerLabel}
+        />
         <Content>
           <ContentHeader title="Time Saver">
             <Tabs value={selectedTab} onChange={handleChange} centered={false}>
@@ -148,7 +162,10 @@ export const TimeSaverPageComponent = () => {
                           <>
                             {selectedTab === 0 && (
                               <Grid container spacing={2}>
-                                <GaugesContainer dates={dates} />
+                                <GaugesContainer
+                                  dates={dates}
+                                  hoursPerDay={hoursPerDay}
+                                />
                                 <Divider variant="fullWidth" />
                                 <Grid xs={6}>
                                   <AllStatsBarChart dates={dates} />